+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 -->
 
 ## [Unreleased]
+## [0.1.0-alpha35] - 2024-10-01
+
+- improve piker: show path if load failed
+
 ## [0.1.0-alpha34] - 2024-10-01
 
 - improve perf: `tagger` will cache the loaded images in memory